backend/requirements.txt
asgiref==3.8.1
bcrypt==4.2.0
blinker==1.8.2
click==8.1.7
dnspython==2.7.0
Flask==3.0.3
Flask-Cors==5.0.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
marshmallow==3.23.0
packaging==24.1
pymongo==4.9.2
python-dotenv==1.0.1
Werkzeug==3.0.4
